==> 00-run.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import matplotlib.font_manager
import os
from os.path import dirname


#step 0 - git pull to stay up to date with the repo on Github
#Step 1 - make sure that there's data on the API --- see emails in the product analytics email account
#Step 2 - run the notebooks noted below on a stat box 

%run '01_data_collection.ipynb'

print("metrics_month:", query_vars['metrics_month'])
print("snapshot:", query_vars['snapshot'])

%run '02_data_handling.ipynb'

# see the report here: https://github.com/wikimedia-research/Editing-movement-metrics/blob/master/03a-report.ipynb
%run '03_analysis.ipynb'

print("data querying and reporting complete")

#1. base
#output/handling/metrics.csv 
#can be added to https://docs.google.com/spreadsheets/d/1jkwqLUsMKrC_lKkum4iEPC7CjUsnYWNCNaNG_-CdR4o/edit#gid=1877720946


#2. metrics 
#output/campaigns_metrics.csv
#can be added to the metrics sheet https://docs.google.com/spreadsheets/d/1rYrJPltouudjiyvqEvz3rb2sG0B9AOZpujjI2LsD4UY/edit#gid=0
